loader/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.9
